django/pyproject.toml
[tool.poetry]
name = "sistemas-embarcados"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Paulo Vanzolini, Lucas Agnez e Rafael Gomes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
uvicorn = "^0.30.6"
pika = "^1.3.2"
djangorestframework = "^3.15.2"
prometheus-client = "^0.20.0"
django-prometheus = "^2.3.1"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
